Security_Scripts/requirements.txt
dnspython==2.6.1
markdown==3.6
pdfminer.six==20231228
python-docx==1.1.2
requests==2.32.3
shodan==1.31.0
python-whois==0.9.4
ipwhois==1.2.0
numpy==2.1.2
scikit-learn==1.5.2
cryptography==43.0.3

